branch update - not bugfree
git-svn-id: https://svn.r-project.org/R/trunk@6098 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/strsignif.c
@@ -93,8 +93,8 @@ void str_signif(char *x, int *n, char **type, int *width, int *digits,
 			strcat(form, "*d");
 		}
 		if (strcmp("integer", *type) == 0)
-		  for (i=0; i < nn; i++)
-			sprintf(result[i], form, wid, (int)((long *)x)[i]);
+		    for (i=0; i < nn; i++)
+			sprintf(result[i], form, wid, ((int *)x)[i]);
 		else
 			error("`type' must be \"integer\" for  \"d\"-format");
 	}
